guard matplotlib import and disable by default
