+++ odetoolbox/mixed_integrator.py
@@ -19,6 +19,8 @@
 # along with NEST.  If not, see <http://www.gnu.org/licenses/>.
 #
 
+from typing import Optional
+
 import logging
 import numpy as np
 import numpy.random
@@ -30,19 +32,9 @@ import time
 
 from .analytic_integrator import AnalyticIntegrator
 from .integrator import Integrator
+from .plot_helper import import_matplotlib
 from .shapes import Shape
 from .sympy_printer import _is_sympy_type
-
-
-try:
-    import matplotlib as mpl
-    mpl.use('Agg')
-    import matplotlib.pyplot as plt
-    INTEGRATOR_DEBUG_PLOT = True
-    INTEGRATOR_DEBUG_PLOT_DIR = "/tmp"
-except ImportError:
-    INTEGRATOR_DEBUG_PLOT = False
-
 
 try:
     import pygsl.odeiv as odeiv
@@ -65,7 +57,7 @@ class MixedIntegrator(Integrator):
     Mixed numeric+analytic integrator. Supply with a result from ODE-toolbox analysis; calculates numeric approximation of the solution.
     """
 
-    def __init__(self, numeric_integrator, system_of_shapes, shapes, analytic_solver_dict=None, parameters=None, spike_times=None, random_seed=123, max_step_size=np.inf, integration_accuracy_abs=1E-6, integration_accuracy_rel=1E-6, sim_time=1., alias_spikes=False):
+    def __init__(self, numeric_integrator, system_of_shapes, shapes, analytic_solver_dict=None, parameters=None, spike_times=None, random_seed=123, max_step_size=np.inf, integration_accuracy_abs=1E-6, integration_accuracy_rel=1E-6, sim_time=1., alias_spikes=False, debug_plot_dir: Optional[str] = None):
         r"""
         :param numeric_integrator: A method from the GSL library for evolving ODEs, e.g. :python:`odeiv.step_rk4`
         :param system_of_shapes: Dynamical system to solve.
@@ -79,11 +71,13 @@ class MixedIntegrator(Integrator):
         :param integration_accuracy_rel: Relative integration accuracy.
         :param sim_time: How long to simulate for.
         :param alias_spikes: Whether to alias spike times to the numerical integration grid. :python:`False` means that precise integration will be used for spike times whenever possible. :python:`True` means that after taking a timestep :math:`dt` and arriving at :math:`t`, spikes from :math:`\langle t - dt, t]` will only be processed at time :math:`t`.
+        :param debug_plot_dir: If given, enable debug plotting to this directory. If enabled, matplotlib is imported and used for plotting.
         """
         super(MixedIntegrator, self).__init__()
 
         assert PYGSL_AVAILABLE
 
+        self._debug_plot_dir = debug_plot_dir
         self.numeric_integrator = numeric_integrator
         self.alias_spikes = alias_spikes
         self.max_step_size = max_step_size
@@ -347,8 +341,8 @@ class MixedIntegrator(Integrator):
             h_log = np.array(h_log)
             y_log = np.array(y_log)
 
-            if INTEGRATOR_DEBUG_PLOT:
-                self.integrator_debug_plot(t_log, h_log, y_log, dir=INTEGRATOR_DEBUG_PLOT_DIR)
+            if self._debug_plot_dir:
+                self.integrator_debug_plot(t_log, h_log, y_log, dir=self._debug_plot_dir)
 
         logging.info("For integrator = " + str(self.numeric_integrator) + ": h_min = " + str(h_min) + ", h_avg = " + str(h_avg) + ", runtime = " + str(runtime))
 
@@ -361,6 +355,9 @@ class MixedIntegrator(Integrator):
 
 
     def integrator_debug_plot(self, t_log, h_log, y_log, dir):
+        mpl, plt = import_matplotlib()
+        assert mpl, "Debug plot was requested for MixedIntegrator, but an exception occurred while importing matplotlib. See the ``debug_plot_dir`` parameter."
+
         if not self.analytic_integrator is None:
             analytic_syms = self.analytic_integrator.get_value(0.).keys()
             analytic_dim = len(analytic_syms)
